requirements.txt
openai>=1.0.0
speechrecognition
pyttsx3
PyAudio
googlesearch-python